refactor(core/molecule): split definitions
